Move form to create a new catalog entry from App.js
