+++ frontend/src/App.js
@@ -65,6 +65,7 @@ import {
     project_thumbnails_ctxt, 
     project_thumbnails
 } from "./thumbnails.js"
+import { CatalogEntryForm } from "./catalog_entry_form.js"
 
 // const server_configs = {
 //     development: {
@@ -134,150 +135,150 @@ import {
 //   document.body.removeChild(element);
 // }
 
-// Component: Form to create new catalog entry (for download or PR)
-// routes: /new-model, /new-datasource
-//
-// The two submit options either download the form data as json, or
-// create a pull request with the new entry on behalf of the user.
-//
-// This Component can be used for both the model and datasource
-// catalogs, through the props.
-//
-// * gh_logged_in - login status (grey out the PR button if not logged in)
-// * schema - json schema object, used to generate the form
-// * catalog_kind - "datasource" or "model",
-function CatalogEntryForm({ gh_logged_in, schema, uiSchema, catalog_kind, catalog_path, download_filename }) {
+// // Component: Form to create new catalog entry (for download or PR)
+// // routes: /new-model, /new-datasource
+// //
+// // The two submit options either download the form data as json, or
+// // create a pull request with the new entry on behalf of the user.
+// //
+// // This Component can be used for both the model and datasource
+// // catalogs, through the props.
+// //
+// // * gh_logged_in - login status (grey out the PR button if not logged in)
+// // * schema - json schema object, used to generate the form
+// // * catalog_kind - "datasource" or "model",
+// function CatalogEntryForm({ gh_logged_in, schema, uiSchema, catalog_kind, catalog_path, download_filename }) {
 
-    // The modal dialogue shows when 'pr_failed' is true.  Separate
-    // state variable (pr_message) for the message, since closing the
-    // modal clears the failure flag, but the message is still visible
-    // briefly
-    const [ pr_message, set_pr_message ] = useState("");
-    const [ pr_failed, set_pr_failed ] = useState(false);
-    const [ pr_loading, set_pr_loading ] = useState(false);
+//     // The modal dialogue shows when 'pr_failed' is true.  Separate
+//     // state variable (pr_message) for the message, since closing the
+//     // modal clears the failure flag, but the message is still visible
+//     // briefly
+//     const [ pr_message, set_pr_message ] = useState("");
+//     const [ pr_failed, set_pr_failed ] = useState(false);
+//     const [ pr_loading, set_pr_loading ] = useState(false);
 
-    // There is a single onSubmit event for both buttons, but can
-    // use the onClick of the button to set this flag and decide
-    // which one
-    let pr_flag;
+//     // There is a single onSubmit event for both buttons, but can
+//     // use the onClick of the button to set this flag and decide
+//     // which one
+//     let pr_flag;
 
-    return (
-        <div className="mb-5">
-            <Modal show={!!pr_failed} onHide={() => set_pr_failed(false)} >
-                <Modal.Header closeButton>
-                    <b>Error opening pull request on GitHub</b>
-                </Modal.Header>
-                <Modal.Body>
-                    The error was:<div className="text-monospace mt-2 col">{pr_message}</div>
-                </Modal.Body>
-            </Modal>
-            <Form onSubmit={
-                      async (new_entry_submission) => {
-                          const entry = new_entry_submission.formData;
-                          if (pr_flag) {
-                              const octokit = new OctokitPRPlugin({
-                                  auth: sessionStorage[GH_TOKEN_KEY]
-                              });
+//     return (
+//         <div className="mb-5">
+//             <Modal show={!!pr_failed} onHide={() => set_pr_failed(false)} >
+//                 <Modal.Header closeButton>
+//                     <b>Error opening pull request on GitHub</b>
+//                 </Modal.Header>
+//                 <Modal.Body>
+//                     The error was:<div className="text-monospace mt-2 col">{pr_message}</div>
+//                 </Modal.Body>
+//             </Modal>
+//             <Form onSubmit={
+//                       async (new_entry_submission) => {
+//                           const entry = new_entry_submission.formData;
+//                           if (pr_flag) {
+//                               const octokit = new OctokitPRPlugin({
+//                                   auth: sessionStorage[GH_TOKEN_KEY]
+//                               });
 
-                              set_pr_loading(true);
+//                               set_pr_loading(true);
 
-                              try {
-                                  // TODO: record gh_username on login
-                                  const {
-                                      data: { login: gh_username, },
-                                  } = await octokit.rest.users.getAuthenticated();
+//                               try {
+//                                   // TODO: record gh_username on login
+//                                   const {
+//                                       data: { login: gh_username, },
+//                                   } = await octokit.rest.users.getAuthenticated();
 
-                                  ///////////
-                                  // Sync user fork
-                                  //
-                                  // This is a workaround to a possible bug in octokit-plugin-create-pull-request
-                                  // triggered when the user has an out of date fork, resulting in a 404.
-                                  // The workaround is just to synchronise the fork.
-                                  //
-                                  // With the workaround there is a theoretical chance that the base repository
-                                  // is updated after the call to mergeUpstream but before createPullRequest, but
-                                  // this can really only be fixed upstream.
+//                                   ///////////
+//                                   // Sync user fork
+//                                   //
+//                                   // This is a workaround to a possible bug in octokit-plugin-create-pull-request
+//                                   // triggered when the user has an out of date fork, resulting in a 404.
+//                                   // The workaround is just to synchronise the fork.
+//                                   //
+//                                   // With the workaround there is a theoretical chance that the base repository
+//                                   // is updated after the call to mergeUpstream but before createPullRequest, but
+//                                   // this can really only be fixed upstream.
 
-                                  const scivision_forks = await octokit.rest.repos.listForks({
-                                      owner: "alan-turing-institute",
-                                      repo: "scivision",
-                                  });
+//                                   const scivision_forks = await octokit.rest.repos.listForks({
+//                                       owner: "alan-turing-institute",
+//                                       repo: "scivision",
+//                                   });
 
-                                  const user_scivision_fork = scivision_forks.data.find(
-                                      (fork) => fork.owner && fork.owner.login == gh_username
-                                  );
+//                                   const user_scivision_fork = scivision_forks.data.find(
+//                                       (fork) => fork.owner && fork.owner.login == gh_username
+//                                   );
 
-                                  // Array.find returns undefined if no matches -- in that case,
-                                  // no need to update anything (octokit.createPullRequest
-                                  // will create a fresh one)
-                                  if (user_scivision_fork !== undefined) {
-                                      await octokit.rest.repos.mergeUpstream({
-                                          owner: gh_username,
-                                          repo: user_scivision_fork.name,
-                                          branch: user_scivision_fork.default_branch,
-                                      });
-                                  }
-                                  //
-                                  ///////////
+//                                   // Array.find returns undefined if no matches -- in that case,
+//                                   // no need to update anything (octokit.createPullRequest
+//                                   // will create a fresh one)
+//                                   if (user_scivision_fork !== undefined) {
+//                                       await octokit.rest.repos.mergeUpstream({
+//                                           owner: gh_username,
+//                                           repo: user_scivision_fork.name,
+//                                           branch: user_scivision_fork.default_branch,
+//                                       });
+//                                   }
+//                                   //
+//                                   ///////////
 
-                                  const response = await octokit.createPullRequest({
-                                      owner: "alan-turing-institute",
-                                      repo: "scivision",
-                                      title: `Add "${entry.name}" to the ${catalog_kind} catalog`,
-                                      body: "This PR was automatically generated by the Scivision frontend.",
-                                      base: "main",
-                                      head: `add-${catalog_kind}-${crypto.randomUUID().toUpperCase()}`,
-                                      update: false,
-                                      forceFork: true,
-                                      changes: [
-                                          {
-                                              files: {
-                                                  [catalog_path]:
-                                                  ({ exists, encoding, content}) => {
-                                                      const content_str = Buffer.from(content, encoding);
-                                                      let cat = JSON.parse(content_str);
-                                                      cat.entries.push(entry);
-                                                      return JSON.stringify(cat, null, 2);
-                                                  }
-                                              },
-                                              commit: `Create entry for "${entry.name}" in the ${catalog_kind} catalog`
-                                          }
-                                      ]
-                                  });
-                                  if (response.status >= 200 && response.status < 300) {
-                                      window.location = response.data.html_url;
-                                  } else {
-                                      throw new Error(response.data);
-                                  }
-                              } catch (e) {
-                                  set_pr_message(e.message);
-                                  set_pr_failed(true);
-                                  set_pr_loading(false);
-                              }
+//                                   const response = await octokit.createPullRequest({
+//                                       owner: "alan-turing-institute",
+//                                       repo: "scivision",
+//                                       title: `Add "${entry.name}" to the ${catalog_kind} catalog`,
+//                                       body: "This PR was automatically generated by the Scivision frontend.",
+//                                       base: "main",
+//                                       head: `add-${catalog_kind}-${crypto.randomUUID().toUpperCase()}`,
+//                                       update: false,
+//                                       forceFork: true,
+//                                       changes: [
+//                                           {
+//                                               files: {
+//                                                   [catalog_path]:
+//                                                   ({ exists, encoding, content}) => {
+//                                                       const content_str = Buffer.from(content, encoding);
+//                                                       let cat = JSON.parse(content_str);
+//                                                       cat.entries.push(entry);
+//                                                       return JSON.stringify(cat, null, 2);
+//                                                   }
+//                                               },
+//                                               commit: `Create entry for "${entry.name}" in the ${catalog_kind} catalog`
+//                                           }
+//                                       ]
+//                                   });
+//                                   if (response.status >= 200 && response.status < 300) {
+//                                       window.location = response.data.html_url;
+//                                   } else {
+//                                       throw new Error(response.data);
+//                                   }
+//                               } catch (e) {
+//                                   set_pr_message(e.message);
+//                                   set_pr_failed(true);
+//                                   set_pr_loading(false);
+//                               }
 
-                          } else {
-                              download(download_filename, JSON.stringify(entry, null, 2));
-                          }
-                      }
-                  }
-                  schema={schema} uiSchema={uiSchema}>
+//                           } else {
+//                               download(download_filename, JSON.stringify(entry, null, 2));
+//                           }
+//                       }
+//                   }
+//                   schema={schema} uiSchema={uiSchema}>
 
-                <button type="submit"
-                        onClick={ () => pr_flag = true }
-                        className="btn btn-primary"
-                        disabled={!gh_logged_in || pr_loading}>
-                    Open Pull Request on GitHub
-                    { pr_loading ? <>&nbsp;<Spinner animation="border" role="status" size="sm"/></> : <></> }
-                    { gh_logged_in ? <></> : <> (login to enable)</> }
-                </button>
-                <button type="submit"
-                        onClick={ () => pr_flag = false }
-                        className="btn btn-link">
-                    Download entry data as json
-                </button>
-            </Form>
-        </div>);
-}
+//                 <button type="submit"
+//                         onClick={ () => pr_flag = true }
+//                         className="btn btn-primary"
+//                         disabled={!gh_logged_in || pr_loading}>
+//                     Open Pull Request on GitHub
+//                     { pr_loading ? <>&nbsp;<Spinner animation="border" role="status" size="sm"/></> : <></> }
+//                     { gh_logged_in ? <></> : <> (login to enable)</> }
+//                 </button>
+//                 <button type="submit"
+//                         onClick={ () => pr_flag = false }
+//                         className="btn btn-link">
+//                     Download entry data as json
+//                 </button>
+//             </Form>
+//         </div>);
+// }
 
 // Obtain the GitHub OAuth token from Scivision backend
 //
